requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
psycopg2-binary==2.9.7
matplotlib==3.7.2
networkx==3.1
Pillow==10.0.0
pandas==2.0.3
seaborn==0.12.2
numpy==1.24.3